+++ lib/EasyPost/Exception/General/EndOfPaginationException.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace EasyPost\Exception\General;
+
+use EasyPost\Constant\Constants;
+
+class EndOfPaginationException extends \Exception
+{
+    /**
+     * EndOfPaginationException constructor.
+     *
+     * @param string $message
+     */
+    public function __construct()
+    {
+        parent::__construct(Constants::END_OF_PAGINATION);
+    }
+}