+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Next Release:
+
+- Add `getNextPage` function that retrieves the next page of results for a paginated collection
+
 ## v6.3.1 (2023-03-22)
 
 - Adds missing `status_detail` property to Tracker